+++ app/__init__.py
@@ -1,20 +1,8 @@
 """Application Init.
 """
-from pytsite import content, events, widget, lang, assetman, browser
+from pytsite import content
 from . import model
 
 # Register application models
 content.register_model('article', model.Article, 'articles')
 content.register_model('page', model.Page, 'pages')
-
-# Inject common arguments on every template render event
-events.listen('pytsite.tpl.render', lambda args: args.update({
-    'sections': list(content.get_sections()),
-    'pages': content.find('page').get(),
-    'language_nav': widget.select.LanguageNav('language-nav'),
-    'search_widget': content.widget.Search('search-article', model='article', title=lang.t('search')),
-}))
-
-browser.include('bootstrap', True)
-assetman.add('css/common.css', True)
-assetman.add('js/common.js', True)