+++ app/themes/default/ep.py
@@ -47,17 +47,11 @@ def article_view(args: dict, inp: dict) -> str:
     e = args['entity']
     exclude_ids = [e.id]
 
-    comments_widget = ''
-    if 'pytsite.disqus' in reg.get('app.autoload'):
-        comments_widget = comments.get_widget('disqus')
-    elif 'pytsite.fb' in reg.get('app.autoload'):
-        comments_widget = comments.get_widget('fb')
-
     args.update({
         'related': _get_articles(exclude_ids, 3, e.section, 'views_count') if e.model == 'article' else [],
         'entity_tags': content.widget.EntityTagCloud('entity-tag-cloud', entity=args['entity']),
         'share_widget': addthis.widget.AddThis('add-this-share') if reg.get('addthis.pub_id') else '',
-        'comments_widget': comments_widget,
+        'comments_widget': comments.get_widget('disqus') if reg.get('disqus.short_name') else '',
         'sidebar': _get_sidebar(exclude_ids),
     })